requirements.txt
# Core web framework
Flask>=2.2.0
Jinja2>=3.1.0
Werkzeug>=2.2.0
itsdangerous>=2.1.0
click>=8.1.0

# NLP and ML
transformers>=4.30.0
torch>=2.6.0
torchvision>=0.18.0
scikit-learn>=1.5.0

# PDF processing
PyMuPDF>=1.23.7

# Translation
googletrans>=4.0.0rc1
httpcore==0.15.0

# Data handling
numpy>=1.26.0
pandas>=2.1.0

# For testing
pytest>=7.0.0

# For environment variable management
python-dotenv>=1.0.0